refactor(popup): Merge PopupReplyV2 and PopupThreadV2 into PopupV2
Combine the create and thread view popups into a single PopupV2
component that routes based on the presence of annotationId. Rename
PopupReplyV2.scss to PopupV2.scss. Add CSS.escape for annotation ID
DOM queries.
