--- src/components/Popups/PopupReplyV2.tsx
@@ -1,131 +1,0 @@
-import React from 'react';
-import { useDispatch } from 'react-redux';
-import { useIntl } from 'react-intl';
-import { BlueprintModernizationProvider } from '@box/blueprint-web';
-import {
-    MentionContextProvider,
-    MessageEditorV2,
-    serializeMentionMarkup,
-} from '@box/threaded-annotations';
-import type { DocumentNodeV2 } from '@box/threaded-annotations';
-import type { JSONContent } from '@tiptap/core';
-import FocusTrap from 'box-ui-elements/es/components/focus-trap/FocusTrap';
-
-import { collaboratorToUserContact } from '../../adapters/threadedAnnotationsAdapters';
-import { fetchCollaboratorsAction } from '../../store/users/actions';
-
-import type { AppThunkDispatch } from '../../store/types';
-
-import createPopper, { PopupReference } from './Popper';
-import type { Instance, Options } from './Popper';
-import messages from './messages';
-import './PopupReplyV2.scss';
-
-export type Props = {
-    onSubmit: (text: string) => void;
-    reference: PopupReference;
-};
-
-const getPopupOptions = (): Partial<Options> => ({
-    modifiers: [
-        { name: 'flip', options: { altAxis: false, fallbackPlacements: ['top', 'left', 'right'] } },
-        { name: 'offset', options: { offset: [0, 15] } },
-        { name: 'preventOverflow', options: { padding: 5 } },
-    ],
-    placement: 'bottom',
-});
-
-const createDocumentNode = (content: JSONContent | null): DocumentNodeV2 => {
-    if (!content) {
-        return { type: 'doc', content: [] };
-    }
-
-    if (content.type === 'doc' && content.content) {
-        return content as DocumentNodeV2;
-    }
-
-    return { type: 'doc', content: [content] } as DocumentNodeV2;
-};
-
-const PopupReplyV2 = ({ onSubmit, reference }: Props): JSX.Element => {
-    const intl = useIntl();
-    const dispatch = useDispatch<AppThunkDispatch>();
-    const popupRef = React.useRef<HTMLDivElement>(null);
-    const popperRef = React.useRef<Instance>();
-    const optionsRef = React.useRef<Partial<Options>>(getPopupOptions());
-
-    React.useEffect(() => {
-        if (popupRef.current) {
-            popperRef.current = createPopper(reference, popupRef.current, optionsRef.current);
-        }
-
-        return () => {
-            popperRef.current?.destroy();
-        };
-    }, [reference]);
-
-    const handleEvent = React.useCallback((event: React.SyntheticEvent) => {
-        event.stopPropagation();
-    }, []);
-
-    const userSelectorProps = React.useMemo(
-        () => ({
-            allowEmptyQuery: true,
-            ariaRoleDescription: intl.formatMessage(messages.ariaLabelMentionSelector),
-            fetchAvatarUrls: async () => ({}),
-            fetchUsers: async (query: string) => {
-                const action = await dispatch(fetchCollaboratorsAction(query));
-                if (fetchCollaboratorsAction.fulfilled.match(action)) {
-                    return action.payload.entries.map(collaboratorToUserContact);
-                }
-                return [];
-            },
-            loadingAriaLabel: intl.formatMessage(messages.ariaLabelMentionLoading),
-        }),
-        [dispatch, intl],
-    );
-
-    const handlePost = React.useCallback(
-        async (content: JSONContent | null): Promise<void> => {
-            const doc = createDocumentNode(content);
-            const { text } = serializeMentionMarkup(doc);
-            onSubmit(text);
-        },
-        [onSubmit],
-    );
-
-    return (
-        <FocusTrap>
-            <div
-                ref={popupRef}
-                aria-label={intl.formatMessage(messages.ariaLabelComment)}
-                className="ba-PopupReplyV2"
-                data-resin-component="popupReplyV2"
-                onClick={handleEvent}
-                onMouseDown={handleEvent}
-                onMouseMove={handleEvent}
-                onMouseUp={handleEvent}
-                onTouchMove={handleEvent}
-                onTouchStart={handleEvent}
-                role="presentation"
-            >
-                <BlueprintModernizationProvider enableModernizedComponents>
-                    <MentionContextProvider value={{
-                        collaborationPopoverProps: {
-                            onSubmit: () => Promise.resolve(),
-                        },
-                    }}>
-                        <MessageEditorV2
-                            isFirstAnnotation
-                            onPost={handlePost}
-                            userSelectorProps={userSelectorProps}
-                        />
-                    </MentionContextProvider>
-                </BlueprintModernizationProvider>
-                <div data-threaded-annotations-portal />
-            </div>
-        </FocusTrap>
-    );
-};
-
-export default PopupReplyV2;
