--- src/components/Popups/__tests__/PopupThreadV2-test.tsx
@@ -1,153 +1,0 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
-import { useDispatch, useSelector } from 'react-redux';
-import PopupThreadV2, { Props } from '../PopupThreadV2';
-
-jest.mock('react-redux', () => ({
-    useDispatch: jest.fn(),
-    useSelector: jest.fn(),
-}));
-
-jest.mock('react-intl', () => ({
-    defineMessages: (msgs: Record<string, unknown>) => msgs,
-    useIntl: () => ({
-        formatMessage: (msg: { defaultMessage: string }) => msg.defaultMessage,
-    }),
-}));
-
-jest.mock('box-ui-elements/es/components/focus-trap/FocusTrap', () => {
-    const ReactMock = jest.requireActual('react');
-    return ({ children }: { children: React.ReactNode }) =>
-        ReactMock.createElement('div', { 'data-testid': 'focus-trap' }, children);
-});
-
-jest.mock('@box/blueprint-web', () => ({
-    BlueprintModernizationProvider: ({ children }: { children: React.ReactNode }) => children,
-}));
-
-jest.mock('@box/threaded-annotations', () => {
-    const ReactMock = jest.requireActual('react');
-    return {
-        MentionContextProvider: ({ children }: { children: React.ReactNode }) =>
-            ReactMock.createElement('div', { 'data-testid': 'mention-context' }, children),
-        ThreadedAnnotationsV2: (props: Record<string, unknown>) =>
-            ReactMock.createElement('div', {
-                'data-testid': 'threaded-annotations-v2',
-                'data-is-annotations': String(props.isAnnotations),
-                'data-messages-count': String((props.messages as unknown[])?.length ?? 0),
-                'data-has-on-delete': String(typeof props.onDelete === 'function'),
-                'data-has-on-post': String(typeof props.onPost === 'function'),
-                'data-has-on-resolve': String(typeof props.onResolve === 'function'),
-                'data-has-on-thread-delete': String(typeof props.onThreadDelete === 'function'),
-                'data-has-on-unresolve': String(typeof props.onUnresolve === 'function'),
-            }),
-        serializeMentionMarkup: jest.fn().mockReturnValue({ hasMention: false, text: '' }),
-    };
-});
-
-jest.mock('../../../store/annotations/actions', () => ({
-    createReplyAction: jest.fn(),
-    deleteAnnotationAction: jest.fn(),
-    setActiveAnnotationIdAction: jest.fn(),
-    updateAnnotationAction: jest.fn(),
-}));
-
-jest.mock('../../../store/users/actions', () => ({
-    fetchCollaboratorsAction: Object.assign(jest.fn(), {
-        fulfilled: { match: jest.fn().mockReturnValue(false) },
-    }),
-}));
-
-const mockUseDispatch = useDispatch as jest.MockedFunction<typeof useDispatch>;
-const mockUseSelector = useSelector as jest.MockedFunction<typeof useSelector>;
-
-describe('PopupThreadV2', () => {
-    const defaults: Props = {
-        annotationId: 'annotation-1',
-        reference: document.createElement('div'),
-    };
-
-    const mockAnnotation = {
-        created_at: '2026-01-01T00:00:00Z',
-        created_by: { id: '100', login: 'test@box.com', name: 'Test User', type: 'user' },
-        description: {
-            created_at: '2026-01-01T00:00:00Z',
-            created_by: { id: '100', login: 'test@box.com', name: 'Test User', type: 'user' },
-            id: 'reply-1',
-            message: 'Hello world',
-            parent: { id: 'annotation-1', type: 'annotation' },
-            type: 'reply',
-        },
-        id: 'annotation-1',
-        modified_at: '2026-01-01T00:00:00Z',
-        modified_by: { id: '100', login: 'test@box.com', name: 'Test User', type: 'user' },
-        permissions: { can_delete: true, can_edit: true, can_resolve: true },
-        replies: [],
-        target: { type: 'point', location: { type: 'page', value: 1 }, x: 0, y: 0 },
-        type: 'annotation',
-    };
-
-    const mockDispatch = jest.fn();
-
-    beforeEach(() => {
-        mockUseDispatch.mockReturnValue(mockDispatch);
-        mockUseSelector.mockReturnValue(mockAnnotation);
-    });
-
-    afterEach(() => {
-        jest.clearAllMocks();
-    });
-
-    test('should render ThreadedAnnotationsV2 inside FocusTrap with MentionContextProvider', () => {
-        render(<PopupThreadV2 {...defaults} />);
-
-        expect(screen.getByTestId('focus-trap')).toBeVisible();
-        expect(screen.getByTestId('mention-context')).toBeVisible();
-        expect(screen.getByTestId('threaded-annotations-v2')).toBeVisible();
-    });
-
-    test('should render with isAnnotations=true', () => {
-        render(<PopupThreadV2 {...defaults} />);
-
-        expect(screen.getByTestId('threaded-annotations-v2').getAttribute('data-is-annotations')).toBe('true');
-    });
-
-    test('should render messages from annotation', () => {
-        render(<PopupThreadV2 {...defaults} />);
-
-        expect(screen.getByTestId('threaded-annotations-v2').getAttribute('data-messages-count')).toBe('1');
-    });
-
-    test('should render with empty messages when annotation is not found', () => {
-        mockUseSelector.mockReturnValue(undefined);
-        render(<PopupThreadV2 {...defaults} />);
-
-        expect(screen.getByTestId('threaded-annotations-v2').getAttribute('data-messages-count')).toBe('0');
-    });
-
-    test('should pass all action callbacks to ThreadedAnnotationsV2', () => {
-        render(<PopupThreadV2 {...defaults} />);
-
-        const thread = screen.getByTestId('threaded-annotations-v2');
-        expect(thread.getAttribute('data-has-on-delete')).toBe('true');
-        expect(thread.getAttribute('data-has-on-post')).toBe('true');
-        expect(thread.getAttribute('data-has-on-resolve')).toBe('true');
-        expect(thread.getAttribute('data-has-on-thread-delete')).toBe('true');
-        expect(thread.getAttribute('data-has-on-unresolve')).toBe('true');
-    });
-
-    test('should set presentation role and aria-label on popup container', () => {
-        render(<PopupThreadV2 {...defaults} />);
-
-        const popup = screen.getByRole('presentation');
-        expect(popup).toHaveAttribute('aria-label', 'Comment');
-        expect(popup).toHaveAttribute('data-resin-component', 'popupThreadV2');
-    });
-
-    test('should render portal container for threaded-annotations popovers', () => {
-        render(<PopupThreadV2 {...defaults} />);
-
-        const portal = screen.getByRole('presentation').querySelector('[data-threaded-annotations-portal]');
-        expect(portal).not.toBeNull();
-    });
-});
